# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 智岗罗盘 API
  version: 1.0.0
  description: |
    多源异构驱动的岗位能力动态演化与人岗匹配系统 API 契约。
    本文件是前后端并行开发的单一事实源（AGENTS.md 铁律一）。
    前端通过 `openapi-typescript` 生成类型，不得自行定义与后端不一致的类型。
    认证方案：Bearer 双 Token（access 30min / refresh 7d）；refresh_token 同时
    写入 httpOnly Cookie（httponly + samesite=lax），供页面刷新后静默恢复会话。
    标注规则：全部端点已实现（M1-M4 全量交付），契约不含里程碑占位标注。
  contact:
    name: 智岗罗盘团队
    url: https://github.com/zhigang-compass
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.zhigang.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/admin/raw/{raw_type}:
    get:
      tags: [admin]
      summary: 课程/论文/社区信号原始数据分页列表（q 关键词匹配快照标题与正文，source 源过滤）
      parameters:
        - name: raw_type
          in: path
          required: true
          schema: { type: string, enum: [course, paper, community] }
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: 关键词（ILIKE 匹配 snapshot 标题与 raw_text）
        - name: source
          in: query
          required: false
          schema: { type: string }
          description: 来源平台过滤（icourse163/coursera/edx/arxiv/github/stackoverflow）
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 分页列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/RawAdminPage'
                  trace_id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/raw/{raw_type}/{raw_id}:
    get:
      tags: [admin]
      summary: 原始数据详情（快照全量 + 正文全文，供编辑回显）
      parameters:
        - name: raw_type
          in: path
          required: true
          schema: { type: string, enum: [course, paper, community] }
        - name: raw_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/RawAdminDetail'
                  trace_id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [admin]
      summary: 编辑原始数据（title/raw_text/source_url；写 AuditLog；编辑随既有 ETL 指纹重算自然生效，图谱节点不联动）
      parameters:
        - name: raw_type
          in: path
          required: true
          schema: { type: string, enum: [course, paper, community] }
        - name: raw_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RawAdminUpdate' }
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data: { type: object }
                  trace_id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin]
      summary: 删除原始数据（写 AuditLog；图谱侧节点为独立备份不联动）
      parameters:
        - name: raw_type
          in: path
          required: true
          schema: { type: string, enum: [course, paper, community] }
        - name: raw_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data: { type: object }
                  trace_id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/skill-aliases:
    get:
      tags: [admin]
      summary: 动态别名表列表（技能别名回写产物，normalize_skill 并查源）
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [pending, approved, rejected], description: 状态过滤（空=全部） }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: 别名表分页列表（variant→standard，approved 行为 normalize_skill 生效源）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/SkillAliasItem' }
                      total: { type: integer }
                      limit: { type: integer }
                      offset: { type: integer }
                  trace_id: { type: string }
...
  /api/v1/admin/skills:
    get:
      tags: [admin]
      summary: 技能治理列表（白名单标准名 ∪ approved 别名标准名；q/分类/白名单/噪声过滤 + 分页）
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 200, description: 关键字（名称或别名模糊匹配） }
        - name: category
          in: query
          required: false
          schema: { type: string, maxLength: 100, description: 分类过滤（模糊） }
        - name: whitelist
          in: query
          required: false
          schema: { type: string, enum: [all, only, exclude], default: all, description: 白名单过滤（only=仅白名单 / exclude=仅非白名单） }
        - name: noise
          in: query
          required: false
          schema: { type: string, enum: [all, only, exclude], default: all, description: 噪声过滤（only=仅噪声 / exclude=仅非噪声） }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: size
          in: query
          required: false
          schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: 技能治理分页列表（data 含 items/total/page/size）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/SkillAdminItem' }
                      total: { type: integer }
                      page: { type: integer }
                      size: { type: integer }
                  trace_id: { type: string }
  /api/v1/admin/skill-aliases/{alias_id}/review:
    post:
      tags: [admin]
      summary: 别名复核（pending → approved/rejected；写 AuditLog；approved 后热刷新动态别名表即时生效）
      parameters:
        - name: alias_id
          in: path
          required: true
          schema: { type: string, description: 别名 UUID }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SkillAliasReviewIn' }
      responses:
        '200':
          description: 复核成功（data 含 id/status/approved）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data: { $ref: '#/components/schemas/SkillAliasReviewOut' }
                  trace_id: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 仅 pending 可复核（当前状态非 pending 返回 409）
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, msg, trace_id]
      properties:
        code:
          type: integer
          description: |
            0=成功，非 0=业务错误码（设计文档 §2.4.7）：4000 参数校验 / 4010 未认证 /
            4011 Token 过期 / 4030 无权限 / 4040 资源不存在 / 4090 资源或状态冲突 /
            4290 请求限流 / 5000 内部错误（含 LLM 能力不可用场景，HTTP 503）/
            5001 Neo4j 查询失败 / 5002 pgvector 查询失败 /
            5003 LLM 超时（HTTP 504）
        msg:
          type: string
        data:
          description: 业务数据（成功时存在）
        trace_id:
          type: string
          description: 链路追踪 ID（uuid4 hex 截断 16 位）
...
    RawAdminItem:
      type: object
      description: 课程/论文/社区原始数据列表项（GET /admin/raw/{raw_type}）
      required: [id, source, source_id]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        source: { type: string }
        source_id: { type: string }
        source_url: { type: string }
        crawled_at: { type: string }
        is_desensitized: { type: boolean }
        text_length: { type: integer, description: "raw_text 长度" }
        updated_at: { type: string }
        extra:
          type: object
          description: "类型特有展示字段（course: quality/institution/skills_count；paper: published/authors_count；community: stars/votes/trend_type；缺失键 null）"
          additionalProperties: true
    RawAdminPage:
      type: object
      description: GET /admin/raw/{raw_type} 响应 data
      required: [total, page, size, items]
      properties:
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/RawAdminItem' }
    RawAdminDetail:
      type: object
      description: GET/PUT /admin/raw/{raw_type}/{raw_id} 详情（快照全量 + 正文全文）
      required: [id, raw_type, source, source_id, raw_text]
      properties:
        id: { type: integer, format: int64 }
        raw_type: { type: string, enum: [course, paper, community] }
        title: { type: string }
        source: { type: string }
        source_id: { type: string }
        source_url: { type: string }
        crawled_at: { type: string }
        is_desensitized: { type: boolean }
        raw_text: { type: string }
        snapshot: { type: object, additionalProperties: true }
    RawAdminUpdate:
      type: object
      description: PUT /admin/raw/{raw_type}/{raw_id} 请求体（仅列出的字段被更新；编辑随既有 ETL 指纹重算自然生效）
      properties:
        title: { type: string }
        source_url: { type: string }
        raw_text: { type: string }
...
    SkillAliasItem:
      type: object
      description: 技能别名回写记录（skill_aliases 行，方案① #559/#561）
      required: [id, variant, standard_name, status, created_at]
      properties:
        id: { type: string, description: UUID }
        variant: { type: string, description: 别名变体（如 .NET Framework / 3D Modeling） }
        standard_name: { type: string, description: 归并目标标准名（known_standard_names 内） }
        status: { type: string, description: pending → approved（normalize_skill 生效）/ rejected }
        proposal_id: { type: string, description: 来源决策记录 id（llm_decision_records） }
        source: { type: string, description: 来源（llm_review） }
        reviewed_by: { type: string, description: 审批人（UUID） }
        review_reason: { type: string, description: 审批理由 }
        confidence: { type: number, nullable: true, description: LLM merge 置信度 }
        applied_to_graph: { type: boolean, description: 图同步进度标记（别名 approve 即生效，此字段预留） }
        created_at: { type: string, description: ISO8601 }
    SkillAdminItem:
      type: object
      description: 技能治理列表项（GET /admin/skills：白名单标准名 ∪ approved 别名标准名聚合）
      required: [name, in_whitelist, is_noise]
      properties:
        name: { type: string, description: 技能标准名 }
        category: { type: string, description: 分类（白名单内取 skill_whitelist.yaml category；白名单外可为空） }
        in_whitelist: { type: boolean, description: 是否在 skill_whitelist.yaml（幻觉防控第三道防线单一事实源） }
        is_noise: { type: boolean, description: 是否命中噪声过滤词 }
        aliases:
          type: array
          description: approved 别名变体列表
          items: { type: string }
...
    SkillAliasReviewIn:
      type: object
      description: POST /admin/skill-aliases/{alias_id}/review 请求体
      required: [approved]
      properties:
        approved: { type: boolean, description: true=通过并写入动态别名表 / false=驳回 }
        reason: { type: string, maxLength: 500, description: 审批理由 }
    SkillAliasReviewOut:
      type: object
      description: 别名复核响应 data
      required: [id, status, approved]
      properties:
        id: { type: string, description: 别名 UUID }
        status: { type: string, description: 复核后状态（approved/rejected） }
        approved: { type: boolean }

    # ============================================================
    # 演化模块端点 data（08-14 契约补全：versions/diff/trends/signals/state-machine/watch）
    # ============================================================
...